Resolve SonarCloud workbench findings
